dom/tests/mochitest/chrome/test_popup_blocker_chrome.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=662519
-->
<window title="Mozilla Bug 662519"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=662519"
     target="_blank">Mozilla Bug 662519</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[

  /** Test for Bug 662519 **/

  SimpleTest.waitForExplicitFinish();

  // We have to enable dom.disable_open_during_load which is disabled
  // by the test harness.
  let prefs = Components.classes["@mozilla.org/preferences-service;1"]
                .getService(Components.interfaces.nsIPrefBranch);
  var gLastDomLoadValue = prefs.getBoolPref("dom.disable_open_during_load");
  prefs.setBoolPref("dom.disable_open_during_load", true);

  let w = window.open("data:text/html,foobar", "", "width=200,height=200");
  ok(w, "The window object shouldn't be null");

  SimpleTest.waitForFocus(function() {
    w.close();
    ok(true, "The popup appeared");

    SimpleTest.waitForFocus(function() {
      let w = window.open("data:text/html,foobar", "", "width=200,height=200");
      ok(w, "The window object shouldn't be null");

      SimpleTest.waitForFocus(function() {
        w.close();

        ok(true, "The popup appeared");
        prefs.setBoolPref("dom.disable_open_during_load", gLastDomLoadValue);
        SimpleTest.finish();
      }, w, false);
    });
  }, w, false);
  ]]>
  </script>
</window>
